+++ be1-go/hub/state/channels.go
@@ -1,4 +1,4 @@
-package hub_state
+package state
 
 import "popstellar/channel"
