+++ be1-go/hub/state/peers.go
@@ -1,4 +1,4 @@
-package hub_state
+package state
 
 import (
 	"golang.org/x/xerrors"
