+++ be1-go/hub/state/threadSafeMap.go
@@ -1,4 +1,4 @@
-package hub_state
+package state
 
 import "sync"
