+++ be1-go/inbox/hub_inbox.go
@@ -1,7 +1,7 @@
 package inbox
 
 import (
-	state "popstellar/hub/standard_hub/hub_state"
+	"popstellar/hub/state"
 	"popstellar/message/query/method/message"
 	"sync"
 )
