+++ be1-go/network/server_test.go
@@ -6,7 +6,6 @@ import (
 	"net/http/httptest"
 	"popstellar/crypto"
 	"popstellar/hub"
-	"popstellar/hub/standard_hub"
 	"testing"
 
 	"github.com/rs/zerolog"
@@ -16,7 +15,7 @@ import (
 func TestServerStartAndShutdown(t *testing.T) {
 	log := zerolog.New(io.Discard)
 
-	h, err := standard_hub.NewHub(crypto.Suite.Point(), "", "", log, nil)
+	h, err := hub.NewHub(crypto.Suite.Point(), "", "", log, nil)
 	require.NoErrorf(t, err, "could not create hub")
 
 	srv := NewServer(h, "", 0, "testsocket", log)
@@ -54,5 +53,5 @@ func TestInfoHandler(t *testing.T) {
 // Utility functions
 
 type fakeHub struct {
-	hub.Hub
+	hub.Huber
 }
